+++ source/Plugins/Vehicle/SteerComponent.cpp
@@ -73,7 +73,7 @@ namespace GASS
 
 		if (name == "Steer")
 		{
-			float max_rad_angle = Math::Deg2Rad(m_MaxSteerAngle);
+			float max_rad_angle = static_cast<float>(Math::Deg2Rad(m_MaxSteerAngle));
 			m_DesiredAngle = value*max_rad_angle;
 		}
 		/*float value = GASS_ANY_CAST<float>(any_mess->GetData("Value"));
